# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Whim 单组织 RBAC API",
    "version": "1.0.0",
    "description": "用户全局唯一，单主部门，多角色、多岗位。数据范围：1全部、2自定义部门、3本部门、4本部门及下级、5本人。岗位不会隐式授予角色。Long 超出 JavaScript 安全整数时序列化为字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8089"
    }
  ],
...
  "paths": {
    "/system/permission": {
      "get": {
        "tags": [
          "权限"
        ],
        "summary": "查询系统权限目录。",
        "operationId": "listAssignablePermissions",
        "description": "操作权限：`system:permission:list`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPermissionList"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/permission/{permissionId}/status": {
      "put": {
        "tags": [
          "权限"
        ],
        "summary": "修改全局权限状态。",
        "operationId": "setPermissionStatus",
        "description": "操作权限：`system:permission:status`。功能授权在服务层校验当前有效账号及权限。",
        "parameters": [
          {
            "name": "permissionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    },
...
    "/system/dept": {
      "get": {
        "tags": [
          "部门"
        ],
        "summary": "查询系统部门节点。",
        "operationId": "listDepartments",
        "description": "操作权限：`system:dept:list`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListDeptVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "部门"
        ],
        "summary": "创建系统部门。",
        "operationId": "createDepartment",
        "description": "操作权限：`system:dept:create`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultId"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeptCreateRequest"
              }
            }
          }
        }
      }
    },
    "/system/dept/{deptId}": {
      "put": {
        "tags": [
          "部门"
        ],
        "summary": "修改系统部门名称与排序。",
        "operationId": "updateDepartment",
        "description": "操作权限：`system:dept:update`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "deptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeptUpdateRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "部门"
        ],
        "summary": "删除没有组织及授权引用的部门。",
        "operationId": "deleteDepartment",
        "description": "操作权限：`system:dept:delete`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "deptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/dept/{deptId}/status": {
      "put": {
        "tags": [
          "部门"
        ],
        "summary": "启用或停用系统部门。",
        "operationId": "setDepartmentStatus",
        "description": "操作权限：`system:dept:status`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "deptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PermissionData": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "menuName": {
            "type": "string"
          },
          "perms": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "menuType": {
            "type": "integer",
            "format": "int32"
          },
          "dataPermission": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "功能权限响应。"
      },
...
      "DeptData": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "parentId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "deptName": {
            "type": "string"
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "组织部门树节点响应。"
      },
...
      "DeptCreateRequest": {
        "type": "object",
        "properties": {
          "parentId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "deptName": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "parentId",
          "deptName",
          "sort"
        ],
        "description": "组织部门创建参数，父部门ID为0表示根部门。"
      },
      "DeptUpdateRequest": {
        "type": "object",
        "properties": {
          "deptName": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "sort": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "deptName",
          "sort"
        ],
        "description": "组织部门基础信息修改参数，不允许隐式移动部门树。"
      },
      "StatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "status"
        ],
        "description": "全局权限状态修改参数。"
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "与 HTTP 状态码一致。"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            ]
          },
          "success": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "code",
          "message",
          "success"
        ]
      },
      "ResultPermissionList": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionData"
            }
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "description": "无返回数据，值为 null。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "ResultId": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "ResultListDeptVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeptData"
            }
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
    }
  }
}